requirements.txt
streamlit
growattServer
plotly
python-dotenv